Generate a scatter plot comparing two gene set enrichment analysis results

Generate a scatter plot of normalized enrichment scores comparing two results for gene set enrichment analysis, e.g. w/wo adjusting for cell proportion differences.

comparegsea_scatter(
  gseares_path1,
  gseares_path2,
  result_names = NULL,
  nes_cutoff = 2,
  pval_cutoff = 0.1,
  pvalflag = TRUE,
  interactive = FALSE
)

Arguments

gseares_path1

path to GSEA output.

gseares_path2

path to a second GSEA output.

result_names

a vector of length 2 indicating the names of the two GSEA results. If NULL, names will be set to c("results1", "results2")

nes_cutoff

normalized enrichment score cutoff to identify enriched gene-sets.

pval_cutoff

p value cutoff to identify enriched gene-sets.

pvalflag

a logical value indicating whether to use adjusted p value in selecting enriched gene-sets. Default to TRUE.

interactive

a logical value indicating whether to generate an interactive plot. Default to FALSE.

Details

this function does not support output from GSEA R implementation