Deconvolution of bulk RNA-seq data

Deconvolution of bulk RNA-seq data based on single-cell reference data. Eight bulk deconvolution methods, along with eight normalization methods and four transformation methods are available.

scdecon(
  bulk,
  ref,
  phenodata,
  filter_ref = TRUE,
  decon_method = c("scaden", "CIBERSORT", "OLS", "nnls", "FARDEEP", "RLR", "MuSiC",
    "SCDC"),
  norm_method = c("none", "LogNormalize", "TMM", "median_ratios", "TPM", "SCTransform",
    "scran", "scater", "Linnorm"),
  trans_method = c("none", "log", "sqrt", "vst"),
  gene_length = NULL,
  lfc_markers = log2(1.5),
  marker_strategy = c("all", "pos_fc", "top_50p_logFC", "top_50p_AveExpr"),
  to_remove = NULL,
  ffpe_artifacts = FALSE,
  prop = NULL,
  cibersortpath = NULL,
  pythonpath = NULL,
  tmpdir = NULL,
  seed = 1234
)

Arguments

bulk

a matrix or data.frame of unnormalizaed & untransformed bulk RNA-seq gene expression values with rows representing genes, columns representing samples

ref

a matrix or data.frame of untransformed scRNA-seq gene expression counts with rows representing genes, columns representing cells. This data will be used to deconvolute provided bulk RNA-seq data.

phenodata

a data.frame with rows representing cells, columns representing cell attributes. It should at least contain the first three columns as:

cell barcodes

cell types

subject ids

filter_ref

logical value indicating whether outlier genes & cells should be removed from the provided reference data. Defaults to TRUE

decon_method

character value specifying the deconvolution method to use. Has to be one of "scaden", "CIBERSORT", "OLS", "nnls", "FARDEEP", "RLR", "MuSiC", "SCDC". See details for more information.

norm_method

character value specifying the normalization method to use for both bulk & reference data. Has to be one of "none","LogNormalize", "TMM", "median_ratios", "TPM", "SCTransform", "scran", "scater", "Linnorm". See details for more information.

trans_method

character value specifying the transformation method to use for both bulk & reference data. Has to be one of "none", "log", "sqrt", "vst". See details for more information.

gene_length

a data.frame with two columns. The first column represents gene names that match with provided bulk data. The second column represents length of each gene. Only applicable when norm_method is selected as "TPM".

lfc_markers

log2 fold change cutoff used to identify marker genes for deconvolution. The option only applicable to marker-gene based approaches, such as CIBERSORT, OLS, nnls, FARDEEP and RLR.

marker_strategy

further strategy in selecting marker genes besides applying the log2 fold change cutoff. Can be chosen from: "all", "pos_fc", "top_50p_logFC" or "top_50p_AveExpr". See details for more information.

to_remove

character value representing the cell type to remove from reference data. Only applicable to simulation experiments in evaluating effect of cell type removal from reference.

ffpe_artifacts

logical value indicating whether to add simulated ffpe artifacts in the bulk data. Only applicable to simulation experiments in evaluating the effect of FFPE artifacts.

prop

a matrix or data.frame of simulated cell proportion values with rows representing cell types, columns representing samples. Only applicable to simulation experiments in evaluating the effect of cell type removal from reference.

cibersortpath

full path to CIBERSORT.R script.

pythonpath

full path to python binary where scaden was installed with.

tmpdir

temporary processing directory for scaden.

seed

random seed used for simulating FFPE artifacts. Only applicable when ffpe_artifacts is set to TRUE.

Value

a list containing two or three elements.

first element

a data.frame of predicted cell-type proportions, with rows representing cell types, columns representing samples.

second element

a data.frame of fitting errors of the algorithm; first column represents sample names, second column represents RMSEs.

optional third element

a data.frame of simulated cell proportion after removing the specified cell_type. Only applicable to simulation experiments.

Details

decon_method should be one of the following:

scaden

a deep learning based method using three multi-layer deep neural nets. To use scaden, you need to firstly install scaden via pip or conda, the provide the python path to pythonpath option.

CIBERSORT

a marker gene based support vectors regression approach. CIBERSOR does not allow redistribution. To use CIBERSORT, you need to request the source code from the authors & provide the path of CIBERSORT.R script to cibersortpath option.

OLS

ordinary least squares.

nnls

non-negative least squares.

FARDEEP

robust regression using least trimmed squares

RLR

robust regression using an M estimator

MuSiC

multi-subject single-cell deconvolution

SCDC

an ENSEMBLE method to integrate deconvolution results from different scRNA-seq datasets

norm_method should be one of the following:

none

no normalization is performed.

LogNormalize

LogNormalize method from seurat.

TMM

TMM method from calcNormFactors function from edgeR.

median_ratios

median ratio method from estimateSizeFactors,DESeqDataSet-method function from DESeq2.

TPM

Transcript per million. TPM has to be chosen if ffpe_artifacts is set to TRUE.

SCTransform

SCTransfrom method from Seurat.

scran

computeSumFactors method from scran.

scater

librarySizeFactors method from scater.

Linnorm

Linnorm method from Linnorm.

trans_method should be one of the following:

none

no transformation is performed.

log2

log2 transformation. 0.1 is added to the data to avoid logarithm of 0s.

sqrt

square root transformation.

vst

varianceStabilizingTransformation method from DESeq2.

marker_strategy should be one of the following:

all

all genes passed fold change threshold will be used.

pos_fc

only genes with positive fold changes will be used.

top_50p_logFC

only genes with top 50 percent positive fold changes will be used.

top_50p_AveExpr

only genes with top 50 percent average expression will be used.

Examples

if (FALSE) {
## generate artificial bulk samples
data(refdata)
phenodata <- data.frame(cellid = colnames(refdata),
                        celltypes = refdata$celltype,
                        subjectid = refdata$subjectid)
bulk_sim <- bulk_generator(ref = GetAssayData(refdata, slot = "data", assay = "SCT"), 
                           phenodata = phenodata, 
                           num_mixtures = 20, 
                           num_mixtures_sprop = 1)

## perform deconvolution based on "OLS" algorithm
decon_res <- scdecon(bulk = bulk_sim[[1]],
                     ref = GetAssayData(refdata, slot = "data", assay = "SCT"),
                     phenodata = phenodata,
                     filter_ref = TRUE,
                     decon_method = "OLS",
                     norm_method = "none",
                     trans_method = "none",
                     marker_strategy = "all")
}